requirements.txt
django==4.2
django_filter==22.1
psycopg==3.1.8
djangorestframework
django-cors-headers
django-debug-toolbar==3.4.0
django-environ